requirements.txt
google-generativeai
ollama
vosk
pyaudio
numpy
pywhispercpp
psutil
pywin32
groq
python-dotenv
rapidfuzz
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-telegram-bot
playwright
apscheduler
reportlab
pytest
